# src/main/resources/openapi.yaml
openapi: "3.0.2"
info:
  title: API Ajmanre
  version: "1.0"
servers:
  - url: http://localhost:{port}
    description: The local API server
    variables:
      port:
        enum:
          - '8443'
          - '443'
        default: '8443'
  - url: https://angular.ajmanre.com:{port}
    description: The local API server
    variables:
      port:
        enum:
          - '8443'
          - '443'
        default: '8443'
...
paths:
...
  /lead:
    post:
      tags:
        - "leadAdd"
      summary: "add lead"
      operationId: "leadPost"
      parameters:
        - in: header
          name: Authorization
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Lead"
      responses:
        "200":
          description: "MessageResponse"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /lead/{page}/{size}:
    get:
      tags:
        - "leadPageSize"
      summary: "lead listing"
      operationId: "leadPageSize"
      parameters:
        - in: path
          name : page
          schema:
            type: integer
            default: 1
          required: true
          description: "page no of the list pagination"
        - in: path
          name : size
          schema:
            type: integer
            default: 5
          required: true
          description: "size of the pagination"
      responses:
        "200":
          description: "lead pagination listing"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadPage"
...
  /lead/{id}:
    put:
      tags:
        - "leadEdit"
      summary: "edit lead"
      operationId: "leadPut"
      parameters:
        - in: path
          name : id
          schema:
            type: string
          required: true
          description: "update lead details by id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Lead"
      responses:
        "200":
          description: "MessageResponse"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    MessageResponse:
      type: "object"
      required:
        - message
      properties:
        message:
          type: "string"
          description: "Message"
        identifier:
          type: "string"
          description: "If identifier has to be in response"
...
    Lead:
      type: "object"
      required:
        - type
        - typeId
        - phoneNo
      properties:
        id:
          type: string
          description: source identifier
        updatedAt:
          type: "string"
          format: "date-time"
        type:
          type: string
        typeId:
          type: string
        name:
          type: string
          description: name of the person who is inquiring
        firstName:
          type: string
          description: first name of the person who is inquiring
        lastName:
          type: string
          description: last name of the person who is inquiring
        email:
          type: string
        phoneNo:
          type: string
    LeadPage:
      type: "object"
      required:
        - page
        - size
        - total
      properties:
        page:
          type: integer
          description: page no of data result in pagination
        size:
          type: integer
          description: no of data elements in page
        data:
          type: array
          items:
            $ref: "#/components/schemas/Lead"
          description: data elements list in page
        total:
          type: integer
          format: int64
          description: total elements in the list
...
security:
  - bearerAuth: []  